feat: connect tasks to users

+++ src/tasks/tasks.service.ts
@@ -5,18 +5,19 @@ import {GetTasksFilterDto} from "./dto/get-tasks-filter.dto";
 import {TaskRepository} from "./task.repository";
 import {InjectRepository} from "@nestjs/typeorm";
 import {Task} from "./task.entity";
+import {User} from "../auth/user.entity";
 
 @Injectable()
 export class TasksService {
     constructor(@InjectRepository(TaskRepository) private taskRepository: TaskRepository) {
     }
 
-    async getTasks(getTasksFilterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(getTasksFilterDto)
+    async getTasks(getTasksFilterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(getTasksFilterDto, user)
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne({id})
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}})
 
         if (!found) {
             throw new NotFoundException(`Task with id(${id}) not found.`)
@@ -25,12 +26,12 @@ export class TasksService {
         return found
     }
 
-    createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto)
+    createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user)
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const found = await this.getTaskById(id)
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const found = await this.getTaskById(id, user)
 
         found.status = status
         await found.save()
@@ -38,8 +39,8 @@ export class TasksService {
         return found
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id)
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id})
 
         if (!result.affected) {
             throw new NotFoundException(`Task with id(${id}) not found.`)
